Added deep traversal unit test. Fixed several bugs of pages throwing errors following NHibernate 1.2 upgrade.

+++ SnCore.Web/App_Code/WebSocialService.cs
@@ -107,7 +107,7 @@ namespace SnCore.WebServices
             using (SnCore.Data.Hibernate.Session.OpenConnection(GetNewConnection()))
             {
                 ISession session = SnCore.Data.Hibernate.Session.Current;
-                return (int) (long) queryoptions.CreateCountQuery(session).UniqueResult();
+                return (int) queryoptions.CreateCountQuery(session).UniqueResult();
             }
         }
